Remove unnecessary bounds on lobby creation
In addition, the input form is fixed and properly makes use of the
in-built validation in browsers.

The bounds were unnecessary and served no real purpose. Users can still
select what they selected before this patch, but also things that were
arbitrarily disallowed. As for the server, having 10000 players in a single
lobby is surely extreme, but shouldn't be much worse than having 10000 players
in separate sessions. The network load for drawing should be O(n),
and the network load for messages is O(n^2), but considering that it's just
text, and each user sends maybe 4 characters per second, even with 10k players
it's only ca. 381 MiB/s worth of messages sent to users.