use inbuild canvas drawing to improve performance
This *massively* improves performance. The lack in performance becomes
really apparent when using touch, as touch events are emitted more
frequently, at least on Firefox.

I don't know if there is any special reason to manually implement
Bresenham and more, but if it is really necessary Webassembly could be
used to improve performance.