Remove base64 hack for cursor
Instead we now use encodeURIComponent, which works for both the fill tool and the normal cursor.
On my previous attempt, I used encodeURI, which seemingly didn't encode enough, causing a silent failure for the fill bucket.
On top of that, we now use ` for strings in these methods, to avoid unnecessary escaping.