Add basic localization
This commit introduces basic localization.
The main HTML pages and the footer have been translation into us english and german for now.
This changes the locale in the HTML and all strings. Number formats and the likes are not localized for now.
There's basic testing and validation.
Some things have also been renamed in the english version.