Improve performance of loading the whole drawing
Previously, we simply executed all `line` and `fill` events.
However, this causes many reads and writes to the canvas context.
By manipulating the canvas data and only putting it back into the canvas
context after all manipulations are done we are saving a lot of
ressources and gain a huge performance improvement.