Refactor homepage
This refactoring is part of the rewrite to preact. Since I have dropped
the rewrite, but really like what I did (in comparison to what was there
before at least), I rewrote it to vanillajs, reusing the CSS and HTML.

This also changes the background for the lobby.

For now, there are inconsistencies between home and lobby.