Remove player socket mutex
This SHOULD not be require anymore, as it was just due to the lack of
thread-safety in gorilla. gws claims to be doing this properly.

While the player could disconnect, causing us to null the socket after
it has been retrieved, this shouldn't really make a difference to
before, as even with the mutex we could run into a disconnected / invalid
connection.