Colors are indices now
Technically this makes the actual colors a client concern. However, it
also allows the client to have a different color palette an easy way.

The main reason for doing this however, was reducing payload size, in
preparation for the drawing gallery, to reduce memory usage when saving
drawings and potentially persisting them.

We might also change this to a binary format, as it'll be much smaller,
given the amount of JSON overhead.

A minimum line event will currently have a size of 70b, while in binary,
it would only have 11 bytes, if we don't microoptimise (less is
possible).