New scoring algorithms
There are now two scoring algorithms from which the user can chose.

The code base has been prepared to technically support more algorithms in the future.

The two algorithms are alterations of the previous algorithm, but with adjusted scaling. By default, the scaling got more lax, while the second variant is more competitive, with an emphasis on speed.